restructure extract-epub to perform a single extraction per invocation, so that we have one per tempdir and no longer need to produce a directory structure. This is important for cases when the path to the zip file is long or the zip file's name contains whitespace, which the unzip utility doesn't handle well in the exdir parameter.

+++ extract-epub.sh
@@ -1,12 +1,17 @@
 #!/usr/bin/env zsh
 
-tempdir=~/.$$
-mkdir $tempdir
-for i in "$@" ; do
-	unzip -d "$tempdir/${i}.d" "$i"
-done
-find "$tempdir" | egrep '\.(xhtml|html|htm)$' | sort | while read x ; do
-	w3m -dump "$x" 
-done
-rm -rf $tempdir
+
+if [[ $# -gt 1 ]] ; then
+	for i in "$@" ; do
+		$0 "$i"
+	done
+else
+	tempdir=~/.$$
+	mkdir "$tempdir"
+	unzip -d "$tempdir" "$1"
+	find "$tempdir" | egrep '\.(xhtml|html|htm)$' | sort | while read x ; do
+		w3m -dump "$x" 
+	done
+	rm -rf $tempdir
+fi
